ui refactor email settings
